requirements.txt
requests>=2.32.5
tqdm>=4.67.1
python-dotenv>=1.0.0
openai>=1.0.0
